backend/requirements.txt
# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# =============================================================================
# Database (PostgreSQL + pgvector)
# =============================================================================
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
pgvector>=0.3.6

# =============================================================================
# LLM & Agent Orchestration
# =============================================================================
langchain>=0.3.14
langchain-openai>=0.3.0
langgraph>=0.2.60
openai>=1.59.0

# Optional: Anthropic provider (install if LLM_PROVIDER=anthropic)
# langchain-anthropic>=0.3.0

# =============================================================================
# Ingestion Pipeline
# =============================================================================
youtube-transcript-api>=1.0.0
yt-dlp>=2024.12.0

# =============================================================================
# Observability
# =============================================================================
structlog>=24.4.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.1
httpx>=0.28.0

# =============================================================================
# Testing
# =============================================================================
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
